+++ app/src/main/java/com/ikholopov/yamblz/weather/weathermobilization/data/dto/weather/Wind.java
@@ -2,5 +2,5 @@ package com.ikholopov.yamblz.weather.weathermobilization.data.dto.weather;
 
 public class Wind {
     float speed;
-    int deg;
+    float deg;
 }
